+++ src/components/LoginForm/LoginForm.jsx
@@ -27,11 +27,11 @@ function LoginForm() {
   return (
     <form className="formPanel" onSubmit={login}>
       <h2>Login</h2>
-      {errors.loginMessage && (
+      {/* {errors.loginMessage && (
         <h3 className="alert" role="alert">
           {errors.loginMessage}
         </h3>
-      )}
+      )} */}
       <div>
         <label htmlFor="username">
           Username:
