+++ src/redux/sagas/user.saga.js
@@ -13,7 +13,7 @@ function* fetchUser() {
     // allow the server session to recognize the user
     // If a user is logged in, this will return their information
     // from the server session (req.user)
-    const response = yield axios.get('/api/user', config);
+    const response = yield axios.get('/api/user/admin', config);
 
     // now that the session has given us a user object
     // with an id and username set the client-side user object to let
